+++ test/libsolidity/syntaxTests/natspec/docstring_author_function.sol
@@ -1,0 +1,6 @@
+contract C {
+    /// @author author
+    function iHaveAuthor() public pure {}
+}
+// ----
+// Warning 9843: (17-35): Documentation tag @author is only allowed on contract definitions. It will be disallowed in 0.7.0.