+++ Changelog.md
@@ -26,12 +26,14 @@ Bugfixes:
 
 
 Language Features:
+ * General: Add unit denomination ``gwei``
 
 
 Compiler Features:
  * NatSpec: Add fields "kind" and "version" to the JSON output.
+ * NatSpec: Inherit tags from unique base if derived function does not provide any.
  * Commandline Interface: Prevent some incompatible commandline options from being used together.
-
+ * NatSpec: Support NatSpec comments on events.
 
 Bugfixes:
  * NatSpec: Do not consider ``////`` and ``/***`` as NatSpec comments.